requirements.txt
# my deps

Django==4.0.3
django-crispy-forms==1.14.0
psycopg2-binary==2.9.3
django-summernote==0.8.20
uwsgi==2.0.20
invoke==1.7.0
django-bootstrap-datepicker-plus==4.0.0
celery==5.2.6
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1
drf-yasg==1.20.0
djangorestframework-simplejwt==5.1.0

# system deps

amqp==5.1.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
bleach==5.0.0
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
kombu==5.2.4
prompt-toolkit==3.0.29
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
#webencodings==0.5.1
